Moved contexts to it's own file so it can easily be replaced with mocks
