+++ highcharts/contexts.js
@@ -1,0 +1,2 @@
+// @ts-nocheck
+export const hcModsContext = require.context('highcharts/modules', true, /\.js$/)
