+++ unit.config.js
@@ -5,7 +5,11 @@ export default {
     'test/unit/module.js',
     'test/unit/components.js'
   ],
-  ignoredByWatcher: ['highcharts/plugin.compiled.js'],
+  ignoredByWatcher: [
+    'highcharts/plugin.compiled.js',
+    'highcharts/contexts.js',
+    'highcharts/contexts.js.bak'
+  ],
   babel: {
     testOptions: {
       plugins: [
